+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 * `OSC 12 ?` to return the cursor color, not the cursor's text color.
 * `OSC 12;#000000` to configure the cursor to use inverted
   foreground/background colors.
+* Call `ioctl(TIOCSCTTY)` on the pts fd in the slave process.
 
 
 ### Security